+++ src/converter.js
@@ -4,6 +4,9 @@ import Big from 'big.js';
 import BN from 'bn.js';
 import {padToEven} from 'ethjs-util';
 
+// use ROUND_HALF_EVEN to align with blockchain rounding
+// Rounds towards nearest neighbour. If equidistant, rounds towards even neighbour.
+// Big.RM = 2;
 const DECIMALS = 18;
 
 /**
@@ -23,14 +26,14 @@ export function convert(num, to, format) {
 
     let result;
     if (to === 'pip') {
-        result = numBig.times(pow).round().toFixed();
+        result = roundToEven(numBig.times(pow), 0);
         if (format === 'hex') {
             return padToEven(new BN(result, 10).toString(16));
         } else {
             return result;
         }
     } else if (to === 'bip') {
-        return numBig.div(pow).toFixed();
+        return roundToEven(numBig.div(pow));
     } else {
         throw new Error('Unknown type');
     }
@@ -87,4 +90,34 @@ export function numberToBig(num) {
     // }
 
     return new Big(num);
+}
+
+/**
+ * Round-to-even
+ * It is used to align with blockchain rounding
+ * @param {Big} value The number to round-to-even
+ * @param {number} decimals The number of decimals in base10 to round the number at.
+ * @return {string}
+ */
+export function roundToEven(value, decimals = DECIMALS) {
+    if (isRounded(value, decimals)) {
+        // already rounded
+        return value.toFixed();
+    }
+    const power = new Big(10).pow(decimals);
+    const result = value.times(power).div(2).round().times(2)
+        .div(power);
+
+    return result.toFixed();
+}
+
+/**
+ *
+ * @param {Big} value
+ * @param {number} decimals
+ * @return {boolean}
+ */
+function isRounded(value, decimals) {
+    const decimalPart = value.toFixed().split('.')[1];
+    return !decimalPart || decimalPart.length <= decimals;
 }