+++ src/variables.js
@@ -1,4 +1,5 @@
 /**
+ * @deprecated
  * Maximum amount of any coin
  * @type {number}
  */
@@ -7,7 +8,7 @@ export const COIN_MAX_AMOUNT = 10 ** 15;
  * Coin's maximum `maxSupply` value (in BIP)
  * @type {number}
  */
-export const COIN_MAX_MAX_SUPPLY = COIN_MAX_AMOUNT;
+export const COIN_MAX_MAX_SUPPLY = 10 ** 15;
 /**
  * Coin's minimum `maxSupply` value (in BIP)
  * @type {number}