+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 - add v2.0 tx types
 - **BREAKING** fee module now designed as constructor, which accepts current network fee values and returns instance with method `getFeeValue`
 - add `payloadLength` param to `getFeeValue`
+- add `roundToEven` method
 - **BREAKING** TX_TYPE.EDIT_COIN_OWNER renamed to TX_TYPE.EDIT_TICKER_OWNER
 - **BREAKING** remove deprecated `numToBig` and default export from converter module, use `numberToBig` and named exports instead
+- **BREAKING** converter now rounds to nearest even
 
 ## 0.20.0 - 2020.12.30
 - **BREAKING** change fee calculation for Create Coin tx (now it depends on unit value)